Map generation simplified, rewritten, and now much MUCH more readable!
